Initial commit
There is nothing special around here now. My Makefile.build library, some
dummy Makefiles set up and stuff. And a new `aesthetic' target to re-format
documents as there are at least two english documents planned: a
specification of the P language in p-lang/LANGUAGE and the communication
protocol specification in proto/PROTOCOL.

The first time ever that I import my projects' .vimrc and templates into a
repository, though. The .vimrc function LoadTemplate() assumes that the file
is created relative to the project root. In fact this is recommended and
must be the best method to operate on the source tree since I do it most of
the time ;-) - the build process also operates only from the project root.

You will have to configure your vim to load this project's .vimrc, of
course.

There is also some source code already and half-ways written version of the
protocol (don't look at it; I just commit in the middle of doing it). This
is what I did until Weisz finally chose to tell me about his decision about
the name ;-) And I want the skeleton to be up now.