Write some hypothetical P functions
The repository got a vim template and vim syntax highlighting support for P
source files. Note that nothing is done in the actual code. These are just
tests to get a grasp of what the language will need. The asm-gen.c and
preproc.c files are empty, IIRC.

p-lang/primes.P contains _lots_ of comments on features. I will tidy these
up and then make a clear plan once the network protocol is on its way.
Interesting how a primality test in the collective could look like, though!

The p-lang/goldbach.P chronologically first and in all rawness imposed the
need on me to make tuples a native data type of P. Most thoughts, however,
are to be found in the big primes.P file, not all of which are compatible or
listed in chronological order.