Begin working on native tuple support
The routines in lib/tuple.c are meant for the P language. There are still
some issues left to be solved like if nested tuples should be allowed or how
this concept could be replaced, should there be a distinction about easy and
complex native data types, etc..

Since the P language indirectly accesses the depot DBMS routines using
tuples, it would be good if the code can equivalently well be used by the
DBMS. We shall see this later. The get_tuple_hash() function is something we
would really like to implement then. Not to mention that it would be so much
faster (if we stay on the local machine) if the backing structure would not
need any conversion between what is manipulated from P and seen by the DMBS.

As said in the depot/README, we will need a good binary search tree for
SHA-1 (or one of the bigger SHA-2 algorithms) hashes. I hope that creating
include/btree.h reminds me of that betimes.